+++ timer.h
@@ -11,16 +11,16 @@ class Timer : public QObject
 public:
     Timer();
     ~Timer();
-    quint32  counterFine = 0;
+    qint32  counterFine = 0;
     void timerStart();
     void timerStop();
 private:
-    quint16 counter = 0;
+    qint32 counter = 0;
     QTimer *timerInstance = nullptr;
 private slots:
     void timer_slot();
 signals:
-    void set_infobar_time_signal(const QString&, const quint16&);
+    void set_infobar_time_signal(const QString&, const qint32&);
 };
 
 #endif // TIMER_H
